DEV-113 DELETE AND REFACTOR old security system.
