+++ src/main/java/com/study/stockmanagementstudycase/auth/model/dto/request/TokenRefreshRequest.java
@@ -1,6 +1,6 @@
-package com.study.stockmanagementstudycase.security.model.dto.response;
+package com.study.stockmanagementstudycase.auth.model.dto.request;
 
-
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,9 +10,9 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-@NoArgsConstructor
 @AllArgsConstructor
-public class AuthenticationResponse {
-    private String accessToken;
+@NoArgsConstructor
+public class TokenRefreshRequest {
+    @NotBlank
     private String refreshToken;
 }
