--- src/main/java/com/study/stockmanagementstudycase/security/config/SecurityConfig.java
@@ -1,74 +1,0 @@
-package com.study.stockmanagementstudycase.security.config;
-
-import com.study.stockmanagementstudycase.security.filter.JWTAuthenticationFilter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.authentication.logout.LogoutHandler;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-@EnableMethodSecurity
-public class SecurityConfig {
-
-    private final JWTAuthenticationFilter jwtAuthenticationFilter;
-    private final AuthenticationProvider authenticationProvider;
-    private final LogoutHandler logoutHandler;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(
-            final HttpSecurity httpSecurity
-    ) throws Exception {
-
-        httpSecurity
-                .csrf(AbstractHttpConfigurer::disable)
-                .cors(customizer -> customizer.configurationSource(corsConfigurationSource()))
-                .authorizeHttpRequests(
-                        httpRequest ->
-                                httpRequest
-                                        .requestMatchers("/api/v1/auth/**").permitAll()
-                                        .requestMatchers("/api/v1/**").authenticated()
-                )
-                .sessionManagement(httpRequest -> httpRequest.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authenticationProvider(authenticationProvider)
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-                .logout(
-                        httpSecurityLogoutConfigurer ->
-                                httpSecurityLogoutConfigurer
-                                        .addLogoutHandler(logoutHandler)
-                                        .logoutUrl("/api/v1/auth/logout")
-                                        .logoutSuccessHandler(
-                                                (request, response, authentication) -> SecurityContextHolder.clearContext()
-                                        )
-
-                );
-
-        return httpSecurity.build();
-    }
-
-
-    private CorsConfigurationSource corsConfigurationSource() {
-        final CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowedOrigins(List.of("*"));
-        configuration.setAllowedMethods(List.of("*"));
-        configuration.setAllowedHeaders(List.of("*"));
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-}
