--- src/main/java/com/study/stockmanagementstudycase/security/model/Token.java
@@ -1,27 +1,0 @@
-package com.study.stockmanagementstudycase.security.model;
-
-import com.study.stockmanagementstudycase.security.model.entity.TokenEntity;
-import com.study.stockmanagementstudycase.security.model.enums.TokenType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * Domain model of {@link TokenEntity}
- */
-@Builder
-@EqualsAndHashCode
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Token {
-    public String id;
-    private String token;
-    private TokenType tokenType = TokenType.BEARER;
-    private boolean revoked;
-    private boolean expired;
-}
