--- src/main/java/com/study/stockmanagementstudycase/security/model/dto/request/AuthenticationRequest.java
@@ -1,15 +1,0 @@
-package com.study.stockmanagementstudycase.security.model.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    private String password;
-}
