--- src/main/java/com/study/stockmanagementstudycase/security/model/dto/request/RegisterRequest.java
@@ -1,17 +1,0 @@
-package com.study.stockmanagementstudycase.security.model.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RegisterRequest {
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-}
