--- src/main/java/com/study/stockmanagementstudycase/security/model/entity/TokenEntity.java
@@ -1,48 +1,0 @@
-package com.study.stockmanagementstudycase.security.model.entity;
-
-import com.study.stockmanagementstudycase.common.model.entity.BaseEntity;
-import com.study.stockmanagementstudycase.security.model.enums.TokenType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@Setter
-@SuperBuilder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "TOKEN")
-public class TokenEntity extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    public String id;
-
-    @Column(unique = true)
-    private String token;
-
-    @Enumerated(EnumType.STRING)
-    private TokenType tokenType = TokenType.BEARER;
-
-    private boolean revoked;
-
-    private boolean expired;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    public UserEntity user;
-}
