--- src/main/java/com/study/stockmanagementstudycase/security/service/JWTService.java
@@ -1,128 +1,0 @@
-package com.study.stockmanagementstudycase.security.service;
-
-import com.study.stockmanagementstudycase.security.model.enums.TokenType;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
-
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-@Service
-public class JWTService {
-
-    @Value("${application.security.jwt.secret-key}")
-    private String secretKey;
-
-    @Value("${application.security.jwt.expiration}")
-    private long jwtExpiration;
-
-    @Value("${application.security.jwt.refresh-token.expiration}")
-    private long refreshExpiration;
-
-
-    private Key getSignInKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    private Claims extractAllClaims(String token) {
-        return Jwts
-                .parserBuilder()
-                .setSigningKey(getSignInKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    public <T> T extractClaim(
-            final String token,
-            final Function<Claims, T> claimsResolver) {
-        final Claims claims = extractAllClaims(token);
-        return claimsResolver.apply(claims);
-    }
-
-    public String extractUsername(
-            final String token
-    ) {
-        return extractClaim(token, Claims::getSubject);
-    }
-
-    public String generateToken(
-            final UserDetails userDetails
-    ) {
-        return generateToken(new HashMap<>(), userDetails);
-    }
-
-    public String generateToken(
-            final Map<String, Object> extraClaims,
-            final UserDetails userDetails
-    ) {
-        return buildToken(extraClaims, userDetails, jwtExpiration);
-    }
-
-    private String buildToken(
-            final Map<String, Object> extraClaims,
-            final UserDetails userDetails,
-            final long expiration
-    ) {
-        return Jwts
-                .builder()
-                .setClaims(extraClaims)
-                .setSubject(userDetails.getUsername())
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + expiration))
-                .signWith(getSignInKey(), SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public String generateRefreshToken(
-            final UserDetails userDetails
-    ) {
-        return buildToken(new HashMap<>(), userDetails, refreshExpiration);
-    }
-
-    public boolean isTokenValid(
-            final String token,
-            final UserDetails userDetails
-    ) {
-        final String username = extractUsername(token);
-        return (
-                username.equals(userDetails.getUsername())
-                        &&
-                        !isTokenExpired(token)
-        );
-    }
-
-    public boolean isTokenExpired(
-            final String token
-    ) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    private Date extractExpiration(
-            final String token
-    ) {
-        return extractClaim(token, Claims::getExpiration);
-    }
-
-    public String getJwtFromHeader(
-            final HttpServletRequest request
-    ) {
-        final String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
-        if (authHeader == null || !authHeader.startsWith(TokenType.BEARER.getValue())) {
-            throw new RuntimeException();
-        }
-        return authHeader.substring(7);
-    }
-}
